requirements.txt
annotated-types==0.7.0
anyio==4.13.0
bidict==0.23.1
blinker==1.9.0
cachelib==0.13.0
certifi==2026.2.25
cffi==2.0.0
charset-normalizer==3.4.7
click==8.3.2
cryptography==46.0.7
distro==1.9.0
Flask==3.1.3
Flask-Session==0.8.0
Flask-SocketIO==5.6.1
google-ai-generativelanguage==0.6.15
google-api-core==2.30.3
google-api-python-client==2.194.0
google-auth==2.49.2
google-auth-httplib2==0.3.1
google-genai==1.73.1
google-generativeai==0.8.6
googleapis-common-protos==1.74.0
grpcio==1.80.0
grpcio-status==1.71.2
h11==0.16.0
httpcore==1.0.9
httplib2==0.31.2
httpx==0.28.1
idna==3.11
itsdangerous==2.2.0
Jinja2==3.1.6
markdown-it-py==4.0.0
MarkupSafe==3.0.3
maturin==1.13.1
mdurl==0.1.2
msgspec==0.21.1
packaging==25.0
proto-plus==1.27.2
protobuf==5.29.6
pyasn1==0.6.3
pyasn1_modules==0.4.2
pycparser==3.0
pydantic==2.13.0
pydantic_core==2.46.0
Pygments==2.18.0
pyparsing==3.3.2
python-engineio==4.13.1
python-socketio==5.16.1
PyYAML==6.0.2
requests==2.33.1
rich==15.0.0
simple-websocket==1.1.0
sniffio==1.3.1
tenacity==9.1.4
tqdm==4.67.3
typing-inspection==0.4.2
typing_extensions==4.15.0
uritemplate==4.2.0
urllib3==2.6.3
websockets==16.0
Werkzeug==3.1.8
wheel==0.46.1
wsproto==1.3.2
